gendoc

+++ examples/cmajor/parser/Attribute.cpp
@@ -3,7 +3,7 @@
 #include <cmajor/lexer/CmajorLexer.hpp>
 #include <cmajor/lexer/CmajorTokens.hpp>
 
-// this file has been automatically generated from 'D:/work/soulng-project/examples/cmajor/parser/Attribute.parser' using soulng parser generator spg version 1.2.0
+// this file has been automatically generated from 'D:/work/soulng-project/examples/cmajor/parser/Attribute.parser' using soulng parser generator spg version 2.0.0
 
 using namespace soulng::unicode;
 using namespace CmajorTokens;
@@ -29,7 +29,7 @@ soulng::parser::Match AttributeParser::Attributes(CmajorLexer& lexer)
             soulng::parser::Match match(false);
             soulng::parser::Match* parentMatch2 = &match;
             {
-                int pos = lexer.GetPos();
+                int64_t pos = lexer.GetPos();
                 soulng::parser::Match match(false);
                 if (*lexer == LBRACKET)
                 {
@@ -50,7 +50,7 @@ soulng::parser::Match AttributeParser::Attributes(CmajorLexer& lexer)
             soulng::parser::Match* parentMatch3 = &match;
             {
                 soulng::parser::Match match(true);
-                int save = lexer.GetPos();
+                int64_t save = lexer.GetPos();
                 soulng::parser::Match* parentMatch4 = &match;
                 {
                     soulng::parser::Match match(false);
@@ -72,7 +72,7 @@ soulng::parser::Match AttributeParser::Attributes(CmajorLexer& lexer)
                                 {
                                     while (true)
                                     {
-                                        int save = lexer.GetPos();
+                                        int64_t save = lexer.GetPos();
                                         {
                                             soulng::parser::Match match(false);
                                             soulng::parser::Match* parentMatch9 = &match;
@@ -136,11 +136,11 @@ soulng::parser::Match AttributeParser::Attributes(CmajorLexer& lexer)
             soulng::parser::Match match(false);
             soulng::parser::Match* parentMatch12 = &match;
             {
-                int pos = lexer.GetPos();
+                int64_t pos = lexer.GetPos();
                 soulng::parser::Match match(true);
                 soulng::parser::Match* parentMatch13 = &match;
                 {
-                    int pos = lexer.GetPos();
+                    int64_t pos = lexer.GetPos();
                     soulng::parser::Match match(false);
                     if (*lexer == RBRACKET)
                     {
@@ -198,7 +198,7 @@ soulng::parser::Match AttributeParser::Attribute(CmajorLexer& lexer, cmajor::ast
     soulng::parser::Match match(false);
     soulng::parser::Match* parentMatch0 = &match;
     {
-        int pos = lexer.GetPos();
+        int64_t pos = lexer.GetPos();
         soulng::parser::Match match(false);
         soulng::parser::Match* parentMatch1 = &match;
         {
@@ -208,7 +208,7 @@ soulng::parser::Match AttributeParser::Attribute(CmajorLexer& lexer, cmajor::ast
                 soulng::parser::Match match(false);
                 soulng::parser::Match* parentMatch3 = &match;
                 {
-                    int pos = lexer.GetPos();
+                    int64_t pos = lexer.GetPos();
                     soulng::lexer::Span span = lexer.GetSpan();
                     soulng::parser::Match match(false);
                     if (*lexer == ID)
@@ -231,7 +231,7 @@ soulng::parser::Match AttributeParser::Attribute(CmajorLexer& lexer, cmajor::ast
                 soulng::parser::Match* parentMatch4 = &match;
                 {
                     soulng::parser::Match match(true);
-                    int save = lexer.GetPos();
+                    int64_t save = lexer.GetPos();
                     soulng::parser::Match* parentMatch5 = &match;
                     {
                         soulng::parser::Match match(false);
@@ -256,7 +256,7 @@ soulng::parser::Match AttributeParser::Attribute(CmajorLexer& lexer, cmajor::ast
                                     soulng::parser::Match match(false);
                                     soulng::parser::Match* parentMatch9 = &match;
                                     {
-                                        int pos = lexer.GetPos();
+                                        int64_t pos = lexer.GetPos();
                                         soulng::lexer::Span span = lexer.GetSpan();
                                         soulng::parser::Match match(false);
                                         if (*lexer == STRINGLIT)
