+++ examples/cmajor/parser/Enumeration.cpp
@@ -7,7 +7,7 @@
 #include <cmajor/lexer/CmajorLexer.hpp>
 #include <cmajor/lexer/CmajorTokens.hpp>
 
-// this file has been automatically generated from 'D:/work/soulng-project/examples/cmajor/parser/Enumeration.parser' using soulng parser generator spg version 1.2.0
+// this file has been automatically generated from 'D:/work/soulng-project/examples/cmajor/parser/Enumeration.parser' using soulng parser generator spg version 2.0.0
 
 using namespace soulng::unicode;
 using namespace cmajor::ast;
@@ -52,7 +52,7 @@ soulng::parser::Match EnumerationParser::EnumType(CmajorLexer& lexer, ParsingCon
                             soulng::parser::Match match(false);
                             soulng::parser::Match* parentMatch6 = &match;
                             {
-                                int pos = lexer.GetPos();
+                                int64_t pos = lexer.GetPos();
                                 soulng::lexer::Span span = lexer.GetSpan();
                                 soulng::parser::Match match = SpecifierParser::Specifiers(lexer);
                                 specifiers.reset(static_cast<soulng::parser::Value<cmajor::ast::Specifiers>*>(match.value));
@@ -89,12 +89,12 @@ soulng::parser::Match EnumerationParser::EnumType(CmajorLexer& lexer, ParsingCon
                             soulng::parser::Match match(false);
                             soulng::parser::Match* parentMatch9 = &match;
                             {
-                                int pos = lexer.GetPos();
+                                int64_t pos = lexer.GetPos();
                                 soulng::lexer::Span span = lexer.GetSpan();
                                 soulng::parser::Match match(true);
                                 soulng::parser::Match* parentMatch10 = &match;
                                 {
-                                    int pos = lexer.GetPos();
+                                    int64_t pos = lexer.GetPos();
                                     soulng::parser::Match match = IdentifierParser::Identifier(lexer);
                                     enumTypeId.reset(static_cast<IdentifierNode*>(match.value));
                                     if (match.hit)
@@ -125,7 +125,7 @@ soulng::parser::Match EnumerationParser::EnumType(CmajorLexer& lexer, ParsingCon
                     soulng::parser::Match* parentMatch11 = &match;
                     {
                         soulng::parser::Match match(true);
-                        int save = lexer.GetPos();
+                        int64_t save = lexer.GetPos();
                         soulng::parser::Match* parentMatch12 = &match;
                         {
                             soulng::parser::Match match(false);
@@ -134,7 +134,7 @@ soulng::parser::Match EnumerationParser::EnumType(CmajorLexer& lexer, ParsingCon
                                 soulng::parser::Match match(false);
                                 soulng::parser::Match* parentMatch14 = &match;
                                 {
-                                    int pos = lexer.GetPos();
+                                    int64_t pos = lexer.GetPos();
                                     soulng::parser::Match match = EnumerationParser::UnderlyingType(lexer, ctx);
                                     underlyingType.reset(static_cast<Node*>(match.value));
                                     if (match.hit)
@@ -168,12 +168,12 @@ soulng::parser::Match EnumerationParser::EnumType(CmajorLexer& lexer, ParsingCon
                     soulng::parser::Match match(false);
                     soulng::parser::Match* parentMatch16 = &match;
                     {
-                        int pos = lexer.GetPos();
+                        int64_t pos = lexer.GetPos();
                         soulng::lexer::Span span = lexer.GetSpan();
                         soulng::parser::Match match(true);
                         soulng::parser::Match* parentMatch17 = &match;
                         {
-                            int pos = lexer.GetPos();
+                            int64_t pos = lexer.GetPos();
                             soulng::parser::Match match(false);
                             if (*lexer == LBRACE)
                             {
@@ -221,12 +221,12 @@ soulng::parser::Match EnumerationParser::EnumType(CmajorLexer& lexer, ParsingCon
             soulng::parser::Match match(false);
             soulng::parser::Match* parentMatch20 = &match;
             {
-                int pos = lexer.GetPos();
+                int64_t pos = lexer.GetPos();
                 soulng::lexer::Span span = lexer.GetSpan();
                 soulng::parser::Match match(true);
                 soulng::parser::Match* parentMatch21 = &match;
                 {
-                    int pos = lexer.GetPos();
+                    int64_t pos = lexer.GetPos();
                     soulng::parser::Match match(false);
                     if (*lexer == RBRACE)
                     {
@@ -301,7 +301,7 @@ soulng::parser::Match EnumerationParser::UnderlyingType(CmajorLexer& lexer, Pars
             soulng::parser::Match match(false);
             soulng::parser::Match* parentMatch2 = &match;
             {
-                int pos = lexer.GetPos();
+                int64_t pos = lexer.GetPos();
                 soulng::parser::Match match = TypeExprParser::TypeExpr(lexer, ctx);
                 type.reset(static_cast<Node*>(match.value));
                 if (match.hit)
@@ -347,11 +347,11 @@ soulng::parser::Match EnumerationParser::EnumConstants(CmajorLexer& lexer, Parsi
         soulng::parser::Match match(false);
         soulng::parser::Match* parentMatch1 = &match;
         {
-            int pos = lexer.GetPos();
+            int64_t pos = lexer.GetPos();
             soulng::parser::Match match(true);
             soulng::parser::Match* parentMatch2 = &match;
             {
-                int pos = lexer.GetPos();
+                int64_t pos = lexer.GetPos();
                 soulng::parser::Match match = EnumerationParser::EnumConstant(lexer, ctx, enumType);
                 enumConstant.reset(static_cast<EnumConstantNode*>(match.value));
                 if (match.hit)
@@ -381,7 +381,7 @@ soulng::parser::Match EnumerationParser::EnumConstants(CmajorLexer& lexer, Parsi
             {
                 while (true)
                 {
-                    int save = lexer.GetPos();
+                    int64_t save = lexer.GetPos();
                     {
                         soulng::parser::Match match(false);
                         soulng::parser::Match* parentMatch5 = &match;
@@ -402,11 +402,11 @@ soulng::parser::Match EnumerationParser::EnumConstants(CmajorLexer& lexer, Parsi
                                 soulng::parser::Match match(false);
                                 soulng::parser::Match* parentMatch7 = &match;
                                 {
-                                    int pos = lexer.GetPos();
+                                    int64_t pos = lexer.GetPos();
                                     soulng::parser::Match match(true);
                                     soulng::parser::Match* parentMatch8 = &match;
                                     {
-                                        int pos = lexer.GetPos();
+                                        int64_t pos = lexer.GetPos();
                                         soulng::parser::Match match = EnumerationParser::EnumConstant(lexer, ctx, enumType);
                                         enumConstant.reset(static_cast<EnumConstantNode*>(match.value));
                                         if (match.hit)
@@ -474,12 +474,12 @@ soulng::parser::Match EnumerationParser::EnumConstant(CmajorLexer& lexer, Parsin
         soulng::parser::Match match(false);
         soulng::parser::Match* parentMatch1 = &match;
         {
-            int pos = lexer.GetPos();
+            int64_t pos = lexer.GetPos();
             soulng::lexer::Span span = lexer.GetSpan();
             soulng::parser::Match match(true);
             soulng::parser::Match* parentMatch2 = &match;
             {
-                int pos = lexer.GetPos();
+                int64_t pos = lexer.GetPos();
                 soulng::parser::Match match = IdentifierParser::Identifier(lexer);
                 constantId.reset(static_cast<IdentifierNode*>(match.value));
                 if (match.hit)
@@ -510,7 +510,7 @@ soulng::parser::Match EnumerationParser::EnumConstant(CmajorLexer& lexer, Parsin
                 soulng::parser::Match match(false);
                 soulng::parser::Match* parentMatch5 = &match;
                 {
-                    int save = lexer.GetPos();
+                    int64_t save = lexer.GetPos();
                     soulng::parser::Match match(false);
                     soulng::parser::Match* parentMatch6 = &match;
                     {
@@ -530,7 +530,7 @@ soulng::parser::Match EnumerationParser::EnumConstant(CmajorLexer& lexer, Parsin
                             soulng::parser::Match match(false);
                             soulng::parser::Match* parentMatch8 = &match;
                             {
-                                int pos = lexer.GetPos();
+                                int64_t pos = lexer.GetPos();
                                 soulng::lexer::Span span = lexer.GetSpan();
                                 soulng::parser::Match match = ExpressionParser::Expression(lexer, ctx);
                                 constantValue.reset(static_cast<Node*>(match.value));
@@ -563,7 +563,7 @@ soulng::parser::Match EnumerationParser::EnumConstant(CmajorLexer& lexer, Parsin
                             soulng::parser::Match match(false);
                             soulng::parser::Match* parentMatch10 = &match;
                             {
-                                int pos = lexer.GetPos();
+                                int64_t pos = lexer.GetPos();
                                 soulng::parser::Match match(true);
                                 if (match.hit)
                                 {
