+++ examples/cmajor/parser/Parameter.cpp
@@ -5,7 +5,7 @@
 #include <cmajor/lexer/CmajorLexer.hpp>
 #include <cmajor/lexer/CmajorTokens.hpp>
 
-// this file has been automatically generated from 'D:/work/soulng-project/examples/cmajor/parser/Parameter.parser' using soulng parser generator spg version 1.2.0
+// this file has been automatically generated from 'D:/work/soulng-project/examples/cmajor/parser/Parameter.parser' using soulng parser generator spg version 2.0.0
 
 using namespace soulng::unicode;
 using namespace cmajor::ast;
@@ -43,7 +43,7 @@ soulng::parser::Match ParameterParser::ParameterList(CmajorLexer& lexer, Parsing
             soulng::parser::Match* parentMatch2 = &match;
             {
                 soulng::parser::Match match(true);
-                int save = lexer.GetPos();
+                int64_t save = lexer.GetPos();
                 soulng::parser::Match* parentMatch3 = &match;
                 {
                     soulng::parser::Match match(false);
@@ -55,7 +55,7 @@ soulng::parser::Match ParameterParser::ParameterList(CmajorLexer& lexer, Parsing
                             soulng::parser::Match match(false);
                             soulng::parser::Match* parentMatch6 = &match;
                             {
-                                int pos = lexer.GetPos();
+                                int64_t pos = lexer.GetPos();
                                 soulng::parser::Match match = ParameterParser::Parameter(lexer, ctx);
                                 param.reset(static_cast<ParameterNode*>(match.value));
                                 if (match.hit)
@@ -76,7 +76,7 @@ soulng::parser::Match ParameterParser::ParameterList(CmajorLexer& lexer, Parsing
                                 {
                                     while (true)
                                     {
-                                        int save = lexer.GetPos();
+                                        int64_t save = lexer.GetPos();
                                         {
                                             soulng::parser::Match match(false);
                                             soulng::parser::Match* parentMatch9 = &match;
@@ -97,7 +97,7 @@ soulng::parser::Match ParameterParser::ParameterList(CmajorLexer& lexer, Parsing
                                                     soulng::parser::Match match(false);
                                                     soulng::parser::Match* parentMatch11 = &match;
                                                     {
-                                                        int pos = lexer.GetPos();
+                                                        int64_t pos = lexer.GetPos();
                                                         soulng::parser::Match match = ParameterParser::Parameter(lexer, ctx);
                                                         param.reset(static_cast<ParameterNode*>(match.value));
                                                         if (match.hit)
@@ -151,7 +151,7 @@ soulng::parser::Match ParameterParser::ParameterList(CmajorLexer& lexer, Parsing
             soulng::parser::Match match(true);
             soulng::parser::Match* parentMatch13 = &match;
             {
-                int pos = lexer.GetPos();
+                int64_t pos = lexer.GetPos();
                 soulng::parser::Match match(false);
                 if (*lexer == RPAREN)
                 {
@@ -198,7 +198,7 @@ soulng::parser::Match ParameterParser::Parameter(CmajorLexer& lexer, ParsingCont
     soulng::parser::Match match(false);
     soulng::parser::Match* parentMatch0 = &match;
     {
-        int pos = lexer.GetPos();
+        int64_t pos = lexer.GetPos();
         soulng::parser::Match match(false);
         soulng::parser::Match* parentMatch1 = &match;
         {
@@ -208,7 +208,7 @@ soulng::parser::Match ParameterParser::Parameter(CmajorLexer& lexer, ParsingCont
                 soulng::parser::Match match(false);
                 soulng::parser::Match* parentMatch3 = &match;
                 {
-                    int pos = lexer.GetPos();
+                    int64_t pos = lexer.GetPos();
                     soulng::lexer::Span span = lexer.GetSpan();
                     soulng::parser::Match match = TypeExprParser::TypeExpr(lexer, ctx);
                     type.reset(static_cast<Node*>(match.value));
@@ -226,7 +226,7 @@ soulng::parser::Match ParameterParser::Parameter(CmajorLexer& lexer, ParsingCont
                 soulng::parser::Match* parentMatch4 = &match;
                 {
                     soulng::parser::Match match(true);
-                    int save = lexer.GetPos();
+                    int64_t save = lexer.GetPos();
                     soulng::parser::Match* parentMatch5 = &match;
                     {
                         soulng::parser::Match match(false);
@@ -235,7 +235,7 @@ soulng::parser::Match ParameterParser::Parameter(CmajorLexer& lexer, ParsingCont
                             soulng::parser::Match match(false);
                             soulng::parser::Match* parentMatch7 = &match;
                             {
-                                int pos = lexer.GetPos();
+                                int64_t pos = lexer.GetPos();
                                 soulng::lexer::Span span = lexer.GetSpan();
                                 soulng::parser::Match match = IdentifierParser::Identifier(lexer);
                                 id.reset(static_cast<IdentifierNode*>(match.value));
